Refactor NODE_USE_SYSTEM_CA implementation: extract platform-specific code
- Extract macOS Security framework code to root_certs_darwin.cpp (333 lines)
- Create root_certs_platform.h header for platform function declarations
- Clean up main root_certs.cpp by removing 300+ lines of macOS-specific code
- Improve code organization and maintainability
- No functional changes - all tests still pass
- Better separation of concerns: each platform has its own file

Files:
- root_certs.cpp: Main certificate loading logic (337 lines, was 600+)
- root_certs_darwin.cpp: macOS Security framework implementation (333 lines)
- root_certs_platform.h: Platform function declarations (17 lines)
