+++ packages/bun-usockets/src/crypto/root_certs.cpp
@@ -8,52 +8,10 @@
 #include "./default_ciphers.h"
 
 // System-specific includes for certificate loading
+#include "./root_certs_platform.h"
 #ifdef _WIN32
 #include <windows.h>
 #include <wincrypt.h>
-#elif defined(__APPLE__)
-#include <dlfcn.h>
-#include <CoreFoundation/CoreFoundation.h>
-// Security framework types and constants - we'll load dynamically
-typedef struct OpaqueSecCertificateRef* SecCertificateRef;
-typedef struct OpaqueSecTrustRef* SecTrustRef;
-typedef struct OpaqueSecPolicyRef* SecPolicyRef;
-typedef int32_t OSStatus;
-typedef uint32_t SecTrustSettingsDomain;
-
-// Security framework constants (from Security/SecBase.h)
-enum {
-    errSecSuccess = 0,
-    errSecItemNotFound = -25300,
-};
-
-// Trust settings domains (from Security/SecTrustSettings.h)  
-enum {
-    kSecTrustSettingsDomainUser = 0,
-    kSecTrustSettingsDomainAdmin = 1,
-    kSecTrustSettingsDomainSystem = 2,
-};
-
-// Trust result types (from Security/SecTrust.h)
-enum {
-    kSecTrustResultInvalid = 0,
-    kSecTrustResultProceed = 1,
-    kSecTrustResultDeny = 3,
-    kSecTrustResultUnspecified = 4,
-    kSecTrustResultRecoverableTrustFailure = 5,
-    kSecTrustResultFatalTrustFailure = 6,
-    kSecTrustResultOtherError = 7,
-};
-
-// Trust settings result (from Security/SecTrustSettings.h)
-enum {
-    kSecTrustSettingsResultInvalid = 0,
-    kSecTrustSettingsResultTrustRoot = 1,
-    kSecTrustSettingsResultTrustAsRoot = 2,
-    kSecTrustSettingsResultDeny = 3,
-    kSecTrustSettingsResultUnspecified = 4,
-};
-
 #else
 // Linux/Unix includes
 #include <dirent.h>
@@ -64,10 +22,8 @@ static const int root_certs_size = sizeof(root_certs) / sizeof(root_certs[0]);
 
 extern "C" void BUN__warn__extra_ca_load_failed(const char* filename, const char* error_msg);
 
-// System certificate loading functions
-static void us_load_system_certificates_linux(STACK_OF(X509) **system_certs);
-static void us_load_system_certificates_macos(STACK_OF(X509) **system_certs);
-static void us_load_system_certificates_windows(STACK_OF(X509) **system_certs);
+// Forward declarations for platform-specific functions
+// (Actual implementations are in platform-specific files)
 
 // Helper function to check if NODE_USE_SYSTEM_CA is enabled
 static bool us_should_use_system_ca() {
@@ -75,224 +31,10 @@ static bool us_should_use_system_ca() {
   return use_system_ca && (strcmp(use_system_ca, "1") == 0 || strcmp(use_system_ca, "true") == 0);
 }
 
-#ifdef __APPLE__
-// Dynamic Security framework loader
-struct SecurityFramework {
-  void* handle;
-  
-  // Core Foundation constants
-  CFStringRef kSecClass;
-  CFStringRef kSecClassCertificate;
-  CFStringRef kSecMatchLimit;
-  CFStringRef kSecMatchLimitAll;
-  CFStringRef kSecReturnRef;
-  CFBooleanRef kCFBooleanTrue;
-  
-  // Security framework function pointers
-  OSStatus (*SecItemCopyMatching)(CFDictionaryRef query, CFTypeRef *result);
-  CFDataRef (*SecCertificateCopyData)(SecCertificateRef certificate);
-  OSStatus (*SecTrustCreateWithCertificates)(CFArrayRef certificates, CFArrayRef policies, SecTrustRef *trust);
-  SecPolicyRef (*SecPolicyCreateSSL)(Boolean server, CFStringRef hostname);
-  Boolean (*SecTrustEvaluateWithError)(SecTrustRef trust, CFErrorRef *error);
-  OSStatus (*SecTrustSettingsCopyTrustSettings)(SecCertificateRef certRef, SecTrustSettingsDomain domain, CFArrayRef *trustSettings);
-  
-  // Constructor/Destructor
-  SecurityFramework() : handle(nullptr), 
-                       kSecClass(nullptr), kSecClassCertificate(nullptr),
-                       kSecMatchLimit(nullptr), kSecMatchLimitAll(nullptr),
-                       kSecReturnRef(nullptr), kCFBooleanTrue(nullptr),
-                       SecItemCopyMatching(nullptr), SecCertificateCopyData(nullptr),
-                       SecTrustCreateWithCertificates(nullptr), SecPolicyCreateSSL(nullptr),
-                       SecTrustEvaluateWithError(nullptr), SecTrustSettingsCopyTrustSettings(nullptr) {}
-  
-  ~SecurityFramework() {
-    if (handle) {
-      dlclose(handle);
-    }
-  }
-  
-  bool load() {
-    if (handle) return true; // Already loaded
-    
-    handle = dlopen("/System/Library/Frameworks/Security.framework/Security", RTLD_LAZY | RTLD_LOCAL);
-    if (!handle) {
-      fprintf(stderr, "Failed to load Security framework: %s\n", dlerror());
-      return false;
-    }
-    
-    // Load Core Foundation constants
-    kSecClass = *(CFStringRef*)dlsym(handle, "kSecClass");
-    kSecClassCertificate = *(CFStringRef*)dlsym(handle, "kSecClassCertificate");
-    kSecMatchLimit = *(CFStringRef*)dlsym(handle, "kSecMatchLimit");
-    kSecMatchLimitAll = *(CFStringRef*)dlsym(handle, "kSecMatchLimitAll");
-    kSecReturnRef = *(CFStringRef*)dlsym(handle, "kSecReturnRef");
-    
-    // Load CoreFoundation constants from system
-    void* cf_handle = dlopen("/System/Library/Frameworks/CoreFoundation.framework/CoreFoundation", RTLD_LAZY | RTLD_LOCAL);
-    if (cf_handle) {
-      kCFBooleanTrue = *(CFBooleanRef*)dlsym(cf_handle, "kCFBooleanTrue");
-      dlclose(cf_handle);
-    }
-    
-    // Load Security framework functions
-    SecItemCopyMatching = (OSStatus (*)(CFDictionaryRef, CFTypeRef*))dlsym(handle, "SecItemCopyMatching");
-    SecCertificateCopyData = (CFDataRef (*)(SecCertificateRef))dlsym(handle, "SecCertificateCopyData");
-    SecTrustCreateWithCertificates = (OSStatus (*)(CFArrayRef, CFArrayRef, SecTrustRef*))dlsym(handle, "SecTrustCreateWithCertificates");
-    SecPolicyCreateSSL = (SecPolicyRef (*)(Boolean, CFStringRef))dlsym(handle, "SecPolicyCreateSSL");
-    SecTrustEvaluateWithError = (Boolean (*)(SecTrustRef, CFErrorRef*))dlsym(handle, "SecTrustEvaluateWithError");
-    SecTrustSettingsCopyTrustSettings = (OSStatus (*)(SecCertificateRef, SecTrustSettingsDomain, CFArrayRef*))dlsym(handle, "SecTrustSettingsCopyTrustSettings");
-    
-    // Check that all required functions are loaded
-    if (!kSecClass || !kSecClassCertificate || !kSecMatchLimit || 
-        !kSecMatchLimitAll || !kSecReturnRef || !kCFBooleanTrue ||
-        !SecItemCopyMatching || !SecCertificateCopyData ||
-        !SecTrustCreateWithCertificates || !SecPolicyCreateSSL ||
-        !SecTrustEvaluateWithError || !SecTrustSettingsCopyTrustSettings) {
-      fprintf(stderr, "Failed to load required Security framework symbols\n");
-      dlclose(handle);
-      handle = nullptr;
-      return false;
-    }
-    
-    return true;
-  }
-};
-
-// Global instance for dynamic loading
-static SecurityFramework* g_security_framework = nullptr;
-
-static SecurityFramework* get_security_framework() {
-  if (!g_security_framework) {
-    g_security_framework = new SecurityFramework();
-    if (!g_security_framework->load()) {
-      delete g_security_framework;
-      g_security_framework = nullptr;
-    }
-  }
-  return g_security_framework;
-}
-
-// Trust status enumeration (mirroring Node.js implementation)
-enum class TrustStatus {
-  TRUSTED,
-  DISTRUSTED,
-  UNSPECIFIED
-};
-
-// Helper function to determine if a certificate is self-issued
-static bool is_certificate_self_issued(X509* cert) {
-  X509_NAME* subject = X509_get_subject_name(cert);
-  X509_NAME* issuer = X509_get_issuer_name(cert);
-  
-  if (!subject || !issuer) {
-    return false;
-  }
-  
-  return X509_NAME_cmp(subject, issuer) == 0;
-}
-
-// Validate certificate trust using Security framework
-static bool is_certificate_trust_valid(SecCertificateRef cert_ref) {
-  SecurityFramework* security = get_security_framework();
-  if (!security) {
-    return false;
-  }
-  
-  SecTrustRef sec_trust = nullptr;
-  CFMutableArrayRef subj_certs = CFArrayCreateMutable(nullptr, 1, &kCFTypeArrayCallBacks);
-  CFArraySetValueAtIndex(subj_certs, 0, cert_ref);
-  
-  SecPolicyRef policy = security->SecPolicyCreateSSL(false, nullptr);
-  CFArrayRef policies = CFArrayCreate(nullptr, (const void**)&policy, 1, &kCFTypeArrayCallBacks);
-  OSStatus ortn = security->SecTrustCreateWithCertificates(subj_certs, policies, &sec_trust);
-  CFRelease(policies);
-  
-  bool result = false;
-  if (ortn == errSecSuccess) {
-    result = security->SecTrustEvaluateWithError(sec_trust, nullptr);
-  }
-  
-  // Clean up
-  if (policy) CFRelease(policy);
-  if (sec_trust) CFRelease(sec_trust);
-  if (subj_certs) CFRelease(subj_certs);
-  
-  return result;
-}
-
-// Check trust settings for policy (simplified version of Node.js implementation)
-static TrustStatus is_trust_settings_trusted_for_policy(CFArrayRef trust_settings, bool is_self_issued) {
-  if (!trust_settings) {
-    return TrustStatus::UNSPECIFIED;
-  }
-  
-  // Empty trust settings array means "always trust this certificate"
-  if (CFArrayGetCount(trust_settings) == 0) {
-    return is_self_issued ? TrustStatus::TRUSTED : TrustStatus::UNSPECIFIED;
-  }
-  
-  // For simplicity, we'll do basic checking here
-  // A full implementation would parse the trust dictionary entries
-  return TrustStatus::UNSPECIFIED;
-}
-
-// Check if certificate is trusted for server auth policy
-static bool is_certificate_trusted_for_policy(X509* cert, SecCertificateRef cert_ref) {
-  SecurityFramework* security = get_security_framework();
-  if (!security) {
-    return false;
-  }
-  
-  bool is_self_issued = is_certificate_self_issued(cert);
-  bool trust_evaluated = false;
-  
-  // Check user trust domain, then admin domain
-  for (const auto& trust_domain : {kSecTrustSettingsDomainUser, kSecTrustSettingsDomainAdmin}) {
-    CFArrayRef trust_settings = nullptr;
-    OSStatus err = security->SecTrustSettingsCopyTrustSettings(cert_ref, trust_domain, &trust_settings);
-    
-    if (err != errSecSuccess && err != errSecItemNotFound) {
-      fprintf(stderr, "Warning: failed to copy trust settings of system certificate: %d\n", err);
-      continue;
-    }
-    
-    if (err == errSecSuccess && trust_settings != nullptr) {
-      TrustStatus result = is_trust_settings_trusted_for_policy(trust_settings, is_self_issued);
-      CFRelease(trust_settings);
-      
-      if (result == TrustStatus::TRUSTED) {
-        return true;
-      } else if (result == TrustStatus::DISTRUSTED) {
-        return false;
-      }
-    }
-    
-    // If no trust settings and we haven't evaluated trust yet, check trust validity
-    if (trust_settings == nullptr && !trust_evaluated) {
-      bool result = is_certificate_trust_valid(cert_ref);
-      if (result) {
-        return true;
-      }
-      trust_evaluated = true;
-    }
-  }
-  
-  return false;
-}
-
-// Cleanup function for Security framework
-static void cleanup_security_framework() {
-  if (g_security_framework) {
-    delete g_security_framework;
-    g_security_framework = nullptr;
-  }
-}
-
-// Use atexit to ensure cleanup on process exit
-static void __attribute__((constructor)) init_security_framework_cleanup() {
-  atexit(cleanup_security_framework);
-}
-#endif
+// Platform-specific system certificate loading implementations are separated:
+// - macOS: root_certs_darwin.cpp (Security framework with dynamic loading)
+// - Windows: us_load_system_certificates_windows() below
+// - Linux/Unix: us_load_system_certificates_linux() below
 
 // This callback is used to avoid the default passphrase callback in OpenSSL
 // which will typically prompt for the passphrase. The prompting is designed
@@ -502,90 +244,9 @@ extern "C" const char *us_get_default_ciphers() {
 }
 
 // Platform-specific implementations for loading system certificates
-#ifdef __APPLE__
-static void us_load_system_certificates_macos(STACK_OF(X509) **system_certs) {
-  *system_certs = sk_X509_new_null();
-  if (*system_certs == NULL) {
-    return;
-  }
 
-  SecurityFramework* security = get_security_framework();
-  if (!security) {
-    fprintf(stderr, "Warning: Unable to load Security framework, skipping system certificates\n");
-    return;
-  }
-
-  // Create search dictionary for certificates
-  CFTypeRef search_keys[] = {
-    security->kSecClass, 
-    security->kSecMatchLimit, 
-    security->kSecReturnRef
-  };
-  CFTypeRef search_values[] = {
-    security->kSecClassCertificate, 
-    security->kSecMatchLimitAll, 
-    security->kCFBooleanTrue
-  };
-  
-  CFDictionaryRef search = CFDictionaryCreate(
-    kCFAllocatorDefault,
-    search_keys,
-    search_values,
-    3,
-    &kCFTypeDictionaryKeyCallBacks,
-    &kCFTypeDictionaryValueCallBacks
-  );
-
-  CFArrayRef certificates = nullptr;
-  OSStatus status = security->SecItemCopyMatching(search, (CFTypeRef*)&certificates);
-  CFRelease(search);
-
-  if (status != errSecSuccess) {
-    fprintf(stderr, "Warning: SecItemCopyMatching failed with status %d\n", status);
-    return;
-  }
-
-  if (!certificates) {
-    fprintf(stderr, "Warning: No certificates found in keychain\n");
-    return;
-  }
-
-  CFIndex count = CFArrayGetCount(certificates);
-  
-  for (CFIndex i = 0; i < count; ++i) {
-    SecCertificateRef cert_ref = (SecCertificateRef)CFArrayGetValueAtIndex(certificates, i);
-    
-    // Get certificate data
-    CFDataRef cert_data = security->SecCertificateCopyData(cert_ref);
-    if (!cert_data) {
-      fprintf(stderr, "Warning: SecCertificateCopyData failed for certificate %ld\n", i);
-      continue;
-    }
-    
-    // Convert to X509
-    const unsigned char* data_ptr = CFDataGetBytePtr(cert_data);
-    long data_len = CFDataGetLength(cert_data);
-    X509* x509_cert = d2i_X509(nullptr, &data_ptr, data_len);
-    CFRelease(cert_data);
-    
-    if (!x509_cert) {
-      fprintf(stderr, "Warning: Failed to parse certificate %ld as X509\n", i);
-      continue;
-    }
-    
-    // Check if certificate is trusted for server authentication
-    if (is_certificate_trusted_for_policy(x509_cert, cert_ref)) {
-      sk_X509_push(*system_certs, x509_cert);
-    } else {
-      X509_free(x509_cert);
-    }
-  }
-  
-  CFRelease(certificates);
-}
-
-#elif defined(_WIN32)
-static void us_load_system_certificates_windows(STACK_OF(X509) **system_certs) {
+#if defined(_WIN32)
+void us_load_system_certificates_windows(STACK_OF(X509) **system_certs) {
   *system_certs = sk_X509_new_null();
   if (*system_certs == NULL) {
     return;
@@ -619,7 +280,7 @@ static void us_load_system_certificates_windows(STACK_OF(X509) **system_certs) {
 
 #else
 // Linux and other Unix-like systems
-static void us_load_system_certificates_linux(STACK_OF(X509) **system_certs) {
+void us_load_system_certificates_linux(STACK_OF(X509) **system_certs) {
   *system_certs = sk_X509_new_null();
   if (*system_certs == NULL) {
     return;
